amelioration mutateur

+++ Code/Elements/elements.h
@@ -2,15 +2,15 @@
 #include "../Moteur/moteur.h"
 
 //Affiche l'élément bouton
-void bouton(const Select select, const int xBasGauche, const int yBasGauche, const int xHautDroite, const int yHautDroite);
+void bouton(const Select select, const float xBasGauche, const float yBasGauche, const float xHautDroite, const float yHautDroite);
 //Affiche l'élément rouage
-void rouage(const Select select, const int x, const int y, const int r);
+void rouage(const Select select, const float x, const float y, const float r);
 //Affiche le plateau
-void plateau(const int teinte);
+void plateau(const float teinte);
 //Affiche un plus (+)
-void plus(const int xBasGauche, const int yBasGauche, const int xHautDroite, const int yHautDroite);
+void plus(const float xBasGauche, const float yBasGauche, const float xHautDroite, const float yHautDroite);
 //Affiche un coeur
-void coeur(const int x, const int y, const float r);
+void coeur(const float x, const float y, const float r);
 //Affiche le menu pause
 void affichePause(const Select *const select);
 //Affiche la zone de texte
